requirements.txt
numpy>=1.24.0
Pillow>=10.0.0
PyQt6>=6.4.0
opencv-python>=4.8.0 